feat: TripService 내 메서드 별로 로그인 유저의 접근권한을 확인하는 로직 추가

+++ Backend/triplet/src/main/java/com/ssafy/triplet/trip/service/TripService.java
@@ -60,7 +60,7 @@ public class TripService {
             Daily daily = new Daily();
             daily.setUser(user);
             daily.setTrip(trip);
-            daily.setDate(startDate); 
+            daily.setDate(startDate);
             daily.setImageUrl(null); // 초기값으로 null 설정
             startDate = startDate.plusDays(1);// 날짜를 하루씩 증가시킴
             dailyRepository.save(daily);
@@ -70,20 +70,30 @@ public class TripService {
 
 
     // 프로젝트 삭제
-    public void removeTrip(Long tripId) {
+    public void removeTrip(User loginUser, Long tripId) {
         //유효성 검증
         //프로젝트 ID를 찾을 수 없을 경우 에러
-        tripRepository.findById(tripId).orElseThrow(() -> new BaseException(ErrorCode.TRIP_ID_NOT_FOUND));
+        Trip trip = tripRepository.findById(tripId).orElseThrow(() -> new BaseException(ErrorCode.TRIP_ID_NOT_FOUND));
+        Long userId = trip.getUser().getUserId();
+        //접근 권한 확인
+        if (!loginUser.getUserId().equals(userId)) {
+            throw new BaseException(ErrorCode.NOT_AUTHORIZED);
+        }
         //삭제
         tripRepository.deleteById(tripId);
     }
 
     // 프로젝트 수정
-    public void editTrip(Long tripId, TripEditDto tripEditDto) {// 여행 시작날짜, 여행 종료날짜, 환율 변경
+    public void editTrip(User loginUser, Long tripId, TripEditDto tripEditDto) {// 여행 시작날짜, 여행 종료날짜, 환율 변경
         //유효성 검증
         tripValidation.checkEditValid(tripEditDto);
         // DB에서 해당 ID의 Trip 찾기
         Trip trip = tripRepository.findById(tripId).orElseThrow(() -> new BaseException(ErrorCode.TRIP_ID_NOT_FOUND));
+        Long userId = trip.getUser().getUserId();
+        //접근 권한 확인
+        if (!loginUser.getUserId().equals(userId)) {
+            throw new BaseException(ErrorCode.NOT_AUTHORIZED);
+        }
         //수정
         trip.setStartDate(tripEditDto.getStartDate());
         trip.setEndDate(tripEditDto.getEndDate());
